+++ packages/renderer/src/model/index.ts
@@ -1,3 +1,5 @@
 export type { RenderEntity } from './RenderEntity'
 export type { RenderWorld } from './RenderWorld'
 export { EMPTY_RENDER_WORLD } from './RenderWorld'
+export type { RenderPosition } from './RenderPosition'
+export { EMPTY_RENDER_POSITION } from './RenderPosition'